.github/scripts/requirements.txt
google-cloud-translate>=3.12.0
lxml>=5.0.0
